use discretized versions of the corridor curves
The main performance hit are the calls to findClosestPoints
(and therefore distanceTo / findOneClosestPoint too). Moreover,
they the main usage for those are bugs-in-waiting, as some
weird corridor shapes would give bad results.

Replace their use by a discretization of the curves and our
own search routine. This yields a significant performance
improvement.