properly handle getting back into the corridor in case we got out
Instead of simply forbidding intersections between the trajectory
and the boundaries, we now check in which direction (inside->outside
or outside->inside) we do so, and accept getting back in.

Additionally, additional costs are added if the system is outside
the corridor, and if it gets near to the border.