follower: fix horizon computation
The main problem with the current horizon computation is that
it could generate horizons that are too far from the current
position.

In these cases, try to find a closer horizon that matches the
desired distance.

Note that we don't "push" the horizon, as too short horizons
usually spells "difficult terrains", and it is therefore a good
thing that the horizon gets closer.