follower: add the direction of the horizon in the set returned by getNextPossibleDirections
It might happen that neither the local corridor direction, nor
the current forward direction lead to the horizon (in which case
the search fails obviously). Add the general direction of the horizon
segment to the values returned by getNextPossibleDirections to mitigate
that effect.

Note that getNextPossibleDirections does not merge the intervals anymore
which leads to too many expansions, but makes sure that the
"non-point-turning" states are made available to the system.